finish compiler with changing data type in sgx

+++ enclave/sort/bitonic.cpp
@@ -1,14 +1,14 @@
 #include "bitonic.h"
 
-void smallBitonicMerge(int *a, int start, int size, int flipped) {
+void smallBitonicMerge(int64_t *a, int64_t start, int64_t size, int64_t flipped) {
   if (size == 1) {
     return;
   } else {
-    int swap = 0;
-    int mid = greatestPowerOfTwoLessThan(size);
-    for (int i = 0; i < size - mid; ++i) {
-      int num1 = a[start + i];
-      int num2 = a[start + mid + i];
+    int64_t swap = 0;
+    int64_t mid = greatestPowerOfTwoLessThan((double)size);
+    for (int64_t i = 0; i < size - mid; ++i) {
+      int64_t num1 = a[start + i];
+      int64_t num2 = a[start + mid + i];
       swap = num1 > num2;
       swap = swap ^ flipped;
       a[start + i] = (!swap * num1) + (swap * num2);
@@ -21,11 +21,11 @@ void smallBitonicMerge(int *a, int start, int size, int flipped) {
 }
 
 //Correct, after testing
-void smallBitonicSort(int *a, int start, int size, int flipped) {
+void smallBitonicSort(int64_t *a, int64_t start, int64_t size, int64_t flipped) {
   if (size <= 1) {
     return ;
   } else {
-    int mid = greatestPowerOfTwoLessThan(size);
+    int64_t mid = greatestPowerOfTwoLessThan((double)size);
     smallBitonicSort(a, start, mid, 1);
     smallBitonicSort(a, start + mid, size - mid, 0);
     smallBitonicMerge(a, start, size, flipped);
@@ -33,31 +33,31 @@ void smallBitonicSort(int *a, int start, int size, int flipped) {
   return;
 }
 
-void bitonicMerge(int structureId, int start, int size, int flipped, int* row1, int* row2) {
+void bitonicMerge(int structureId, int64_t start, int64_t size, int64_t flipped, int64_t* row1, int64_t* row2) {
   if (size < 1) {
     return ;
   } else if (size < MEM_IN_ENCLAVE) {
-    int *trustedMemory = (int*)oe_malloc(size * BLOCK_DATA_SIZE * sizeof(int));
-    for (int i = 0; i < size; ++i) {
+    int64_t *trustedMemory = (int64_t*)oe_malloc(size * BLOCK_DATA_SIZE * sizeof(int64_t));
+    for (int64_t i = 0; i < size; ++i) {
       opOneLinearScanBlock((start + i) * BLOCK_DATA_SIZE, &trustedMemory[i * BLOCK_DATA_SIZE], BLOCK_DATA_SIZE, structureId, 0, 0);
     }
     smallBitonicMerge(trustedMemory, 0, size * BLOCK_DATA_SIZE, flipped);
-    for (int i = 0; i < size; ++i) {
+    for (int64_t i = 0; i < size; ++i) {
       opOneLinearScanBlock((start + i) * BLOCK_DATA_SIZE, &trustedMemory[i * BLOCK_DATA_SIZE], BLOCK_DATA_SIZE, structureId, 1, 0);
     }
     oe_free(trustedMemory);
   } else {
-    int swap = 0;
-    int mid = greatestPowerOfTwoLessThan(size);
-    for (int i = 0; i < size - mid; ++i) {
+    int64_t swap = 0;
+    int64_t mid = greatestPowerOfTwoLessThan((double)size);
+    for (int64_t i = 0; i < size - mid; ++i) {
       opOneLinearScanBlock((start + i) * BLOCK_DATA_SIZE, row1, BLOCK_DATA_SIZE, structureId, 0, 0);
       opOneLinearScanBlock((start + mid + i) * BLOCK_DATA_SIZE, row2, BLOCK_DATA_SIZE, structureId, 0, 0);
-      int num1 = row1[0], num2 = row2[0];
+      int64_t num1 = row1[0], num2 = row2[0];
       swap = num1 > num2;
       swap = swap ^ flipped;
-      for (int j = 0; j < BLOCK_DATA_SIZE; ++j) {
-        int v1 = row1[j];
-        int v2 = row2[j];
+      for (int64_t j = 0; j < BLOCK_DATA_SIZE; ++j) {
+        int64_t v1 = row1[j];
+        int64_t v2 = row2[j];
         row1[j] = (!swap * v1) + (swap * v2);
         row2[j] = (swap * v1) + (!swap * v2);
       }
@@ -70,22 +70,22 @@ void bitonicMerge(int structureId, int start, int size, int flipped, int* row1, 
   return;
 }
 
-void bitonicSort(int structureId, int start, int size, int flipped, int* row1, int* row2) {
+void bitonicSort(int structureId, int64_t start, int64_t size, int64_t flipped, int64_t* row1, int64_t* row2) {
   if (size < 1) {
     return;
   } else if (size < MEM_IN_ENCLAVE) {
-    int *trustedMemory = (int*)oe_malloc(size * BLOCK_DATA_SIZE * sizeof(int));
-    for (int i = 0; i < size; ++i) {
+    int64_t *trustedMemory = (int64_t*)oe_malloc(size * BLOCK_DATA_SIZE * sizeof(int64_t));
+    for (int64_t i = 0; i < size; ++i) {
       opOneLinearScanBlock((start + i) * BLOCK_DATA_SIZE, &trustedMemory[i * BLOCK_DATA_SIZE], BLOCK_DATA_SIZE, structureId, 0, 0);
     }
     smallBitonicSort(trustedMemory, 0, size * BLOCK_DATA_SIZE, flipped);
     // write back
-    for (int i = 0; i < size; ++i) {
+    for (int64_t i = 0; i < size; ++i) {
       opOneLinearScanBlock((start + i) * BLOCK_DATA_SIZE, &trustedMemory[i * BLOCK_DATA_SIZE], BLOCK_DATA_SIZE, structureId, 1, 0);
     }
     oe_free(trustedMemory);
   } else {
-    int mid = greatestPowerOfTwoLessThan(size);
+    int64_t mid = greatestPowerOfTwoLessThan((double)size);
     bitonicSort(structureId, start, mid, 1, row1, row2);
     bitonicSort(structureId, start + mid, size - mid, 0, row1, row2);
     bitonicMerge(structureId, start, size, flipped, row1, row2);
